chore: rename `compiler/internal/get` to `.../modget`
