+++ compiler/internal/modget/get.go
@@ -1,4 +1,4 @@
-package get
+package modget
 
 import (
 	"fmt"
